api-service/personal-service/requirements.txt
fastapi
uvicorn
pydantic
pydantic[email,typing_extensions,dotenv]
python-multipart
requests
gunicorn
uvloop
httptools
pyjwt
cryptography
pymongo